chore: code cleanup — dedup CENC setup, fix clippy correctness lint, tidy lints
- player: extract setup_track_crypto() helper, collapsing the two near-identical
  video/audio CENC tenc-parse + key-resolve blocks into one (~56 -> ~43 lines).
- audio renderer: silence clippy::never_loop on the single-variant command drain
  loop (kept as a loop for forward-compat; documented with a comment).
- mp4 parser: rename apped_hevc_header -> append_hevc_header (typo) and drop a
  redundant clone; update the ffmpeg_hw call site/comment.
- abr: derive Default for AbrStrategy/AbrVideoProfile instead of hand-rolled impls.
- assorted clippy tidies (no behaviour change): saturating_sub, clamp, split_once?,
  unwrap_or_default, elided lifetimes, redundant closures, -scale_x form, &mut [u8].
- .gitignore: ignore local run artifacts (stdout.log, stderr.log, video_*.mp4).

Verified: cargo check -p player and -p app clean (exit 0); 76 lib tests pass;
clippy warnings 35 -> 16 with the deny-by-default correctness error resolved.

+++ player/src/decoders/ffmpeg_audio.rs
@@ -173,7 +173,7 @@ impl AudioDecoder for FfmpegAudioDecoder {
         let mut frame = ffmpeg_next::util::frame::Audio::empty();
         match decoder.receive_frame(&mut frame) {
             Ok(()) => {
-                let pts_ms = frame.pts().unwrap_or(0) as i64;
+                let pts_ms = frame.pts().unwrap_or(0);
 
                 // Authoritative input format comes from the frame itself.
                 // For EAC-3 in particular, the manifest hint is unreliable —
